Closes #358: Read and merge doc-project and project-org pairs
Improved comments and Preconditions checks. Renaming Test classes.
