+++ iis-wf/iis-wf-affmatching/src/main/java/eu/dnetlib/iis/wf/affmatching/hint/DocumentOrganizationBuilder.java
@@ -5,14 +5,16 @@ import java.io.Serializable;
 import org.apache.spark.api.java.JavaPairRDD;
 import org.apache.spark.api.java.JavaRDD;
 
+import com.google.common.base.Preconditions;
+
 import eu.dnetlib.iis.wf.affmatching.model.DocumentOrganization;
 import eu.dnetlib.iis.wf.affmatching.model.DocumentProject;
 import eu.dnetlib.iis.wf.affmatching.model.ProjectOrganization;
 import scala.Tuple2;
 
 /**
- * Creates {@link DocumentOrganization} objects from document project and
- * project organization datastores.
+ * Creates {@link DocumentOrganization} objects based on {@link DocumentProject} and
+ * {@link ProjectOrganization} datastores.
  * 
  * @author mhorst
  *
@@ -29,10 +31,13 @@ public class DocumentOrganizationBuilder implements Serializable {
 	 * 
 	 * @param docProjRDD {@link DocumentProject} relations
 	 * @param projOrgRDD {@link ProjectOrganization} relations
-	 * @param docProjConfidenceLevelThreshold document project relation confidence level threshold
+	 * @param docProjConfidenceLevelThreshold document project relation confidence level threshold, 
+	 * confidence level check is skipped when this parameter is set to null
 	 */
 	public JavaRDD<DocumentOrganization> build(JavaRDD<DocumentProject> docProjRDD,
 			JavaRDD<ProjectOrganization> projOrgRDD, Float docProjConfidenceLevelThreshold) {
+		Preconditions.checkNotNull(docProjRDD);
+		Preconditions.checkNotNull(projOrgRDD);
 		JavaPairRDD<String, DocumentProject> projIdToDocProj = docProjRDD
 				.mapToPair(docProj -> new Tuple2<String, DocumentProject>(docProj.getProjectId(), docProj))
 				.filter(docProj -> (docProjConfidenceLevelThreshold == null
