+++ iis-wf/iis-wf-affmatching/src/test/java/eu/dnetlib/iis/wf/affmatching/read/ProjectOrganizationConverterTest.java
@@ -4,17 +4,17 @@ import static org.junit.Assert.assertEquals;
 
 import org.junit.Test;
 
-import eu.dnetlib.iis.referenceextraction.project.schemas.DocumentToProject;
-import eu.dnetlib.iis.wf.affmatching.model.DocumentProject;
+import eu.dnetlib.iis.importer.schemas.ProjectToOrganization;
+import eu.dnetlib.iis.wf.affmatching.model.ProjectOrganization;
 
 public class ProjectOrganizationConverterTest {
 
-	private DocumentProjectConverter converter = new DocumentProjectConverter();
+	private ProjectOrganizationConverter converter = new ProjectOrganizationConverter();
 
-	private final String docId = "docId";
 	private final String projId = "projId";
-	private final Float confidenceLevel = 0.9f;
-
+	private final String orgId = "orgId";
+	
+	
 	// ------------------------ TESTS --------------------------
 
 	@Test(expected = NullPointerException.class)
@@ -23,46 +23,30 @@ public class ProjectOrganizationConverterTest {
 	}
 
 	@Test(expected = IllegalArgumentException.class)
-	public void convert_blank_document_id() {
-		converter.convert(new DocumentToProject(" ", projId, 1f));
+	public void convert_blank_organization_id() {
+		converter.convert(new ProjectToOrganization(projId, " "));
 	}
 
 	@Test(expected = IllegalArgumentException.class)
 	public void convert_blank_project_id() {
-		converter.convert(new DocumentToProject(docId, " ", 1f));
+		converter.convert(new ProjectToOrganization(" ", orgId));
 	}
 
 	@Test(expected = IllegalArgumentException.class)
-	public void convert_out_of_right_range_confidence_level() {
-		converter.convert(new DocumentToProject(docId, projId, 2f));
-	}
-
-	@Test(expected = IllegalArgumentException.class)
-	public void convert_out_of_left_range_confidence_level() {
-		converter.convert(new DocumentToProject(docId, projId, -1f));
-	}
-
-	@Test(expected = IllegalArgumentException.class)
-	public void convert_null_document_id() {
-		converter.convert(new DocumentToProject(null, projId, confidenceLevel));
+	public void convert_null_organization_id() {
+		converter.convert(new ProjectToOrganization(projId, null));
 	}
 
 	@Test(expected = IllegalArgumentException.class)
 	public void convert_null_project_id() {
-		converter.convert(new DocumentToProject(docId, null, confidenceLevel));
-	}
-
-	@Test(expected = IllegalArgumentException.class)
-	public void convert_null_confidence_level() {
-		converter.convert(new DocumentToProject(docId, projId, null));
+		converter.convert(new ProjectToOrganization(null, orgId));
 	}
 
 	@Test
 	public void convert() {
-		DocumentProject result = converter.convert(new DocumentToProject(docId, projId, confidenceLevel));
-		assertEquals(docId, result.getDocumentId());
+		ProjectOrganization result = converter.convert(new ProjectToOrganization(projId, orgId));
 		assertEquals(projId, result.getProjectId());
-		assertEquals(confidenceLevel, result.getConfidenceLevel());
+		assertEquals(orgId, result.getOrganizationId());
 	}
 
 }
